# internal/controller/http/v1/response/pages.go
package response

import sharedresp "github.com/QING520-MAKER/Xiaoqing-blog-api/internal/controller/http/shared"

type PostSummaryPage = sharedresp.Page[PostSummary]
type NotificationDetailPage = sharedresp.Page[NotificationDetail]